test(ios): harden final computer deletion flow
